+++ src/client/qcc/cli/__init__.py
@@ -7,7 +7,7 @@ from typing_extensions import Annotated
 
 from ..client import QCC
 from ..conf import settings
-from . import auth, manage, session, static, status
+from . import auth, manage, session, static, status, tariff
 
 app = typer.Typer(name="qcc", no_args_is_help=True)
 app.add_typer(auth.app)
@@ -15,6 +15,7 @@ app.add_typer(manage.app)
 app.add_typer(static.app)
 app.add_typer(status.app)
 app.add_typer(session.app)
+app.add_typer(tariff.app)
 
 
 @app.callback()